pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-mental-health-detection"
version = "0.1.0"
description = "AI-based mental health detection system (text/audio emotion + prediction + chatbot + recommendations + Streamlit dashboard)."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Your Name" }]
dependencies = [
  "numpy>=1.24",
  "pandas>=2.0",
  "scikit-learn>=1.3",
  "nltk>=3.8",
  "regex>=2023.12.25",
  "pydantic>=2.6",
  "PyYAML>=6.0.1",
  "streamlit>=1.31",
  "librosa>=0.10.1",
  "soundfile>=0.12.1",
]

[project.optional-dependencies]
dl = [
  "torch>=2.2",
  "transformers>=4.39",
]
dev = [
  "pytest>=8.0",
  "ruff>=0.3",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
